fixed various functions with new model'
