+++ MetaCocktailsSwiftData.xcodeproj/project.xcworkspace/xcuserdata/matthew.hunt.xcuserdatad/Bookmarks/bookmarks.plist
@@ -12,13 +12,13 @@
 					<key>EndingColumnNumber</key>
 					<string>0</string>
 					<key>EndingLineNumber</key>
-					<string>823</string>
+					<string>815</string>
 					<key>StartingColumnNumber</key>
 					<string>0</string>
 					<key>StartingLineNumber</key>
-					<string>822</string>
+					<string>814</string>
 					<key>Timestamp</key>
-					<string>724543236.147634</string>
+					<string>725307852.601451</string>
 				</dict>
 				<key>rebasable-url</key>
 				<dict>
@@ -34,16 +34,16 @@
 			<key>text-context</key>
 			<dict>
 				<key>focused</key>
-				<string>    }
+				<string>        }
 </string>
 				<key>leading</key>
-				<string>        self.preferenceType = .style
-        self.matchesCurrentSearch = true
+				<string>            self.preferenceType = .style
+            self.matchesCurrentSearch = true
 </string>
 				<key>trailing</key>
-				<string>    init(for texture: Texture) {
-        self.name = texture.rawValue
-        self.isTexture = true
+				<string>        
+        init(for texture: Texture) {
+            self.name = texture.rawValue
 </string>
 			</dict>
 			<key>title</key>
@@ -57,15 +57,15 @@
 				<key>location-parameters</key>
 				<dict>
 					<key>EndingColumnNumber</key>
-					<string>25</string>
+					<string>0</string>
 					<key>EndingLineNumber</key>
 					<string>110</string>
 					<key>StartingColumnNumber</key>
-					<string>11</string>
+					<string>0</string>
 					<key>StartingLineNumber</key>
-					<string>110</string>
+					<string>109</string>
 					<key>Timestamp</key>
-					<string>724541005.5812449</string>
+					<string>725307739.642665</string>
 				</dict>
 				<key>rebasable-url</key>
 				<dict>
@@ -81,15 +81,16 @@
 			<key>text-context</key>
 			<dict>
 				<key>focused</key>
-				<string>func createCom</string>
+				<string>static func createComponentArray() -&gt;  [CocktailComponent] {
+</string>
 				<key>leading</key>
 				<string>        getFilteredCocktails()
     }
-    static </string>
+    </string>
 				<key>trailing</key>
-				<string>ponentArray() -&gt;  [CocktailComponent] {
-        
+				<string>        
         var array = [CocktailComponent]()
+        let convertedArray: [CocktailComponent] = generateBoozeCocktailComponents()
 </string>
 			</dict>
 			<key>title</key>
