+++ MetaCocktailsSwiftData/Model/Components/Ingredient.swift
@@ -8,21 +8,18 @@
 import Foundation
 import SwiftData
 
-protocol Ingredient: Codable {
-    var isBooze: Bool { get }
-    var name: String { get }
-}
-
 class CocktailIngredient: Codable, Hashable {
 
-    let ingredient: any Ingredient
+    let id: UUID
+    let ingredient: IngredientType
     let value: Double
     let unit: MeasurementUnit
     
-    init(ingredient: Ingredient, value: Double, unit: MeasurementUnit = .fluidOunces) {
+    init(ingredient: IngredientType, value: Double, unit: MeasurementUnit = .fluidOunces) {
         self.ingredient = ingredient
         self.value = value
         self.unit = unit
+        self.id = UUID()
     }
     
     func localizedVolumetricString(location: Location) -> String {
@@ -34,70 +31,18 @@ class CocktailIngredient: Codable, Hashable {
         }
     }
     
-    // Decodable conformance
-    
-    private enum CodingKeys: CodingKey {
-        case value
-        case unit
-        case ingredient
-    }
-    
-    required init(from decoder: Decoder) throws {
-        let values = try decoder.container(keyedBy: CodingKeys.self)
-
-        value = try values.decode(Double.self, forKey: .value)
-        unit = try values.decode(MeasurementUnit.self, forKey: .unit)
-        
-        do {
-            let boozeObject  = try values.decode(Booze.self, forKey: .ingredient)
-            ingredient = boozeObject
-        }
-        catch {
-            let nonAlcoholicObject = try values.decode(NonAlcoholic.self, forKey: .ingredient)
-            ingredient = nonAlcoholicObject
-        }
-    }
-    
-    // Encodable conformance
-    
-    func encode(to encoder: Encoder) throws {
-        var container = encoder.container(keyedBy: CodingKeys.self)
-        
-        try container.encode(value, forKey: .value)
-        try container.encode(unit, forKey: .unit)
-        try container.encode(ingredient, forKey: .ingredient)
-
-    }
-    
-    // Hashable and Equatable conformance
+    // Equatable + Hashable Conformance
     
     static func == (lhs: CocktailIngredient, rhs: CocktailIngredient) -> Bool {
-        lhs.ingredient.name > rhs.ingredient.name
+        lhs.id > rhs.id
     }
     
     func hash(into hasher: inout Hasher) {
-        hasher.combine(ingredient.name)
+        hasher.combine(id)
     }
 }
 
 
 
 
-enum MeasurementUnit: String, Codable {
-    case grams          = "grams"
-    case fluidOunces    = "oz"
-    case teaspoons      = "tsp"
-    case disc           = "disc, with flesh"
-    case limeQuarters   = "lime, quartered"
-    case slices         = "slices"
-    case sliceOf        = "slice of"
-    case dash           = "dash"
-    case dashes         = "dashes"
-    case berries        = "berries"
-    case leaves         = "leaves"
-    case ml             = "ml"
-    case barspoon       = "Barspoon"
-    case pinch          = "Pinch"
-    case whole          = "Whole (except the shell)"
-    case sprays         = "Sprays in the glass (glass rinse)"
-}
+
