+++ MetaCocktailsSwiftData/Model/SearchTags/Tags.swift
@@ -9,11 +9,62 @@ import Foundation
 import SwiftUI
 
 struct Tags: Codable {
-   
     var flavors: [Flavor]?
     var profiles: [Profile]?
     var textures: [Texture]?
     var styles: [Style]?
     var booze: [Booze]?
     
+    mutating func merge(with tags: Tags) {
+        var flavors = self.flavors ?? [Flavor]()
+        var profiles = self.profiles ?? [Profile]()
+        var booze = self.booze ?? [Booze]()
+        var textures = self.textures ?? [Texture]()
+        var styles = self.styles ?? [Style]()
+        
+        //self.whatever should be able to stay nil
+        
+        if let newFlavors = tags.flavors {
+            if self.flavors == nil {
+                self.flavors = tags.flavors
+            } else {
+                self.flavors?.append(contentsOf: newFlavors)
+            }
+        }
+        
+        if let newProfiles = tags.profiles {
+            if self.profiles == nil {
+                self.profiles = tags.profiles
+            } else {
+                self.profiles?.append(contentsOf: newProfiles)
+            }
+        }
+        
+        if let newBooze = tags.booze {
+            if self.booze == nil {
+                self.booze = tags.booze
+            } else {
+                self.booze?.append(contentsOf: newBooze)
+            }
+        }
+        
+        if let newTextures = tags.textures  {
+            if self.textures == nil {
+                self.textures = tags.textures
+            } else {
+                self.textures?.append(contentsOf: newTextures)
+            }
+            
+            if let newStyles = tags.styles {
+                if self.styles == nil {
+                    self.styles = tags.styles
+                } else {
+                    self.styles?.append(contentsOf: newStyles)
+                }
+            }
+        }
+        
+    }
 }
+
+
