+++ MetaCocktailsSwiftData/Views/Search View/SearchCriteriaViewModel.swift
@@ -9,7 +9,7 @@ import SwiftUI
 
 
 final class SearchCriteriaViewModel: ObservableObject {
-
+    
     
     @Published var searchText: String = ""
     @Published var cocktailComponents = createComponentArray().sorted(by: { $0.name < $1.name })
@@ -17,34 +17,33 @@ final class SearchCriteriaViewModel: ObservableObject {
     @Published var preferredCount = 0
     @Published var sections = [ResultViewSectionData]()
     @Published var isLoading = true
-    @Published var boozeCategories = { var array: [BoozeCategory] = [BoozeCategory]()
-        for booze in BoozeCategory.allCases {
-            array.append(booze)
-        }
-        return array
-    }()
-        
-    
+    @Published var boozeCategories = ["Agave", "Brandy", "Gin", "OtherAlcohol", "Rum", "Vodka"]
     
     static func generateBoozeCocktailComponents()  -> [CocktailComponent] {
         
-        let boozeCategoryArray: [BoozeProtocol] = [Agave(), Brandy(), Gin(), OtherBooze(), Rum(), Vodka(), Whiskies(), Liqueur(), FortifiedWine(), Wine(), Bitters(), Amari()]
-        var boozeArray: [Booze] =  [Booze]()
-        var arrayOfComponents: [CocktailComponent] = [CocktailComponent]()
         
-        for category in boozeCategoryArray {
-            for expression in category.expressions {
-                let boozeObject = Booze(name: expression, boozeCategory: category.type, isBooze: true)
-                boozeArray.append(boozeObject)
-            }
-        }
-        for booze in boozeArray {
-            arrayOfComponents.append(CocktailComponent(for: booze))
-        }
         
-        return arrayOfComponents
     }
-
+    
+    
+    
+    //        let boozeCategoryArray: [BoozeProtocol] = [Agave(), Brandy(), Gin(), OtherBooze(), Rum(), Vodka(), Whiskies(), Liqueur(), FortifiedWine(), Wine(), Bitters(), Amari()]
+    //        var boozeArray: [Booze] =  [Booze]()
+    //        var arrayOfComponents: [CocktailComponent] = [CocktailComponent]()
+    //
+    //        for category in boozeCategoryArray {
+    //            for expression in category.expressions {
+    //                let boozeObject = Booze(name: expression, boozeCategory: category.type, isBooze: true)
+    //                boozeArray.append(boozeObject)
+    //            }
+    //        }
+    //        for booze in boozeArray {
+    //            arrayOfComponents.append(CocktailComponent(for: booze))
+    //        }
+    //
+    //        return arrayOfComponents
+    
+    
     func matchAllTheThings() {
         // if searchText is empty, show everything again
         
@@ -70,7 +69,7 @@ final class SearchCriteriaViewModel: ObservableObject {
     
     func selectedPreferredIngredients() -> [CocktailComponent] {
         self.cocktailComponents.filter({ $0.isPreferred })
-
+        
     }
     
     func selectedUnwantedIngredients() -> [CocktailComponent] {
@@ -93,12 +92,12 @@ final class SearchCriteriaViewModel: ObservableObject {
     
     func removePreference(for component: CocktailComponent) {
         // When we click a green bubble to remove a preferred tag, we change the data and the UI updates.
-
+        
         preferredCount -= 1
         let cocktailComponent = cocktailComponents.filter ({ $0.id == component.id })
         cocktailComponent.first?.isPreferred = false
-            getFilteredCocktails()
-
+        getFilteredCocktails()
+        
     }
     
     func removeUnwanted(for component: CocktailComponent) {
@@ -117,7 +116,7 @@ final class SearchCriteriaViewModel: ObservableObject {
         for flavor in Flavor.allCases {
             array.append(CocktailComponent(for: flavor))
         }
-
+        
         for profile in Profile.allCases {
             array.append(CocktailComponent(for: profile))
         }
@@ -130,7 +129,7 @@ final class SearchCriteriaViewModel: ObservableObject {
         for component in convertedArray {
             array.append(component)
         }
-
+        
         return array
     }
     
@@ -153,7 +152,7 @@ final class SearchCriteriaViewModel: ObservableObject {
         let flavorCount = selectedPreferredIngredients().filter({ $0.isFlavor }).count
         
         var matchedCount = 0
-
+        
         // Make an array of selected bases called PREFERREDBASES. If no bases are selected, include all.
         
         var preferredBases: [CocktailComponent] = preferredArray.filter({ $0.isSpirit })
@@ -198,129 +197,129 @@ final class SearchCriteriaViewModel: ObservableObject {
         // For every cocktail we got, rip out the bases from each one, and if one of those bases matches one from the PREFERREDBASES array, create a NEW array called MATCHEDBASESCOCKTAILS and throw 'em in.
         // A better way to do this would be to filter the CocktailListViewModel().cocktails.CompileTags().bases by those contained in PREFERREDBASES - but that's a TBD.
         for cocktail in CocktailListViewModel().cocktails {
-//            if let agave = cocktail.CompileTags().agave {
-//                for agave in agave {
-//                    for preferred in preferredBases {
-//                        if agave.rawValue.lowercased() == preferred.name.lowercased() && cocktail != allCocktailsThatMatchBySpirit.last {
-//                            allCocktailsThatMatchBySpirit.append(cocktail)
-//                        }
-//                    }
-//                    
-//                }
-//            }
-//            if let brandies = cocktail.CompileTags().brandy {
-//                for brandy in brandies {
-//                    for preferred in preferredBases {
-//                        if brandy.rawValue.lowercased() == preferred.name.lowercased() && cocktail != allCocktailsThatMatchBySpirit.last {
-//                            allCocktailsThatMatchBySpirit.append(cocktail)
-//                        }
-//                    }
-//                }
-//            }
-//            if let amari = cocktail.CompileTags().amari {
-//                for amaro in amari {
-//                    for preferred in preferredBases {
-//                        if amaro.rawValue.lowercased() == preferred.name.lowercased() && cocktail != allCocktailsThatMatchBySpirit.last {
-//                            allCocktailsThatMatchBySpirit.append(cocktail)
-//                        }
-//                    }
-//                    
-//                }
-//            }
-//            if let bitters = cocktail.CompileTags().bitters{
-//                for bitter in bitters {
-//                    for preferred in preferredBases {
-//                        if bitter.rawValue.lowercased() == preferred.name.lowercased() && cocktail != allCocktailsThatMatchBySpirit.last {
-//                            allCocktailsThatMatchBySpirit.append(cocktail)
-//                        }
-//                    }
-//                    
-//                }
-//            }
-//            if let fortifiedWines = cocktail.CompileTags().fortifiedWine {
-//                for fortifiedWine in fortifiedWines {
-//                    for preferred in preferredBases {
-//                        if fortifiedWine.rawValue.lowercased() == preferred.name.lowercased() && cocktail != allCocktailsThatMatchBySpirit.last {
-//                            allCocktailsThatMatchBySpirit.append(cocktail)
-//                        }
-//                    }
-//                    
-//                }
-//            }
-//            if let gins = cocktail.CompileTags().gin {
-//                for gin in gins {
-//                    for preferred in preferredBases {
-//                        if gin.rawValue.lowercased() == preferred.name.lowercased() && cocktail != allCocktailsThatMatchBySpirit.last {
-//                            allCocktailsThatMatchBySpirit.append(cocktail)
-//                        }
-//                    }
-//                    
-//                }
-//            }
-//            if let liqueurs = cocktail.CompileTags().liqueur {
-//                for liqueur in liqueurs {
-//                    for preferred in preferredBases {
-//                        if liqueur.rawValue.lowercased() == preferred.name.lowercased() && cocktail != allCocktailsThatMatchBySpirit.last {
-//                            allCocktailsThatMatchBySpirit.append(cocktail)
-//                        }
-//                    }
-//                    
-//                }
-//            }
-//            if let others = cocktail.CompileTags().other {
-//                for other in others {
-//                    for preferred in preferredBases {
-//                        if other.rawValue.lowercased() == preferred.name.lowercased() && cocktail != allCocktailsThatMatchBySpirit.last {
-//                            allCocktailsThatMatchBySpirit.append(cocktail)
-//                        }
-//                    }
-//                    
-//                }
-//            }
-//            if let rums = cocktail.CompileTags().rum {
-//                for rum in rums {
-//                    for preferred in preferredBases {
-//                        if rum.rawValue.lowercased() == preferred.name.lowercased() && cocktail != allCocktailsThatMatchBySpirit.last {
-//                            allCocktailsThatMatchBySpirit.append(cocktail)
-//                        }
-//                    }
-//                    
-//                }
-//            }
-//            if let vodkas = cocktail.CompileTags().vodka {
-//                for vodka in vodkas {
-//                    for preferred in preferredBases {
-//                        if vodka.rawValue.lowercased() == preferred.name.lowercased() && cocktail != allCocktailsThatMatchBySpirit.last {
-//                            allCocktailsThatMatchBySpirit.append(cocktail)
-//                        }
-//                    }
-//                    
-//                }
-//            }
-//            if let whiskies = cocktail.CompileTags().whiskies {
-//                for whisky in whiskies {
-//                    for preferred in preferredBases {
-//                        if whisky.rawValue.lowercased() == preferred.name.lowercased() && cocktail != allCocktailsThatMatchBySpirit.last {
-//                            allCocktailsThatMatchBySpirit.append(cocktail)
-//                        }
-//                    }
-//                    
-//                }
-//            }
-//            if let wines = cocktail.CompileTags().wine {
-//                for wine in wines {
-//                    for preferred in preferredBases {
-//                        if wine.rawValue.lowercased() == preferred.name.lowercased() && cocktail != allCocktailsThatMatchBySpirit.last {
-//                            allCocktailsThatMatchBySpirit.append(cocktail)
-//                        }
-//                    }
-//                    
-//                }
-//            }
+            //            if let agave = cocktail.CompileTags().agave {
+            //                for agave in agave {
+            //                    for preferred in preferredBases {
+            //                        if agave.rawValue.lowercased() == preferred.name.lowercased() && cocktail != allCocktailsThatMatchBySpirit.last {
+            //                            allCocktailsThatMatchBySpirit.append(cocktail)
+            //                        }
+            //                    }
+            //
+            //                }
+            //            }
+            //            if let brandies = cocktail.CompileTags().brandy {
+            //                for brandy in brandies {
+            //                    for preferred in preferredBases {
+            //                        if brandy.rawValue.lowercased() == preferred.name.lowercased() && cocktail != allCocktailsThatMatchBySpirit.last {
+            //                            allCocktailsThatMatchBySpirit.append(cocktail)
+            //                        }
+            //                    }
+            //                }
+            //            }
+            //            if let amari = cocktail.CompileTags().amari {
+            //                for amaro in amari {
+            //                    for preferred in preferredBases {
+            //                        if amaro.rawValue.lowercased() == preferred.name.lowercased() && cocktail != allCocktailsThatMatchBySpirit.last {
+            //                            allCocktailsThatMatchBySpirit.append(cocktail)
+            //                        }
+            //                    }
+            //
+            //                }
+            //            }
+            //            if let bitters = cocktail.CompileTags().bitters{
+            //                for bitter in bitters {
+            //                    for preferred in preferredBases {
+            //                        if bitter.rawValue.lowercased() == preferred.name.lowercased() && cocktail != allCocktailsThatMatchBySpirit.last {
+            //                            allCocktailsThatMatchBySpirit.append(cocktail)
+            //                        }
+            //                    }
+            //
+            //                }
+            //            }
+            //            if let fortifiedWines = cocktail.CompileTags().fortifiedWine {
+            //                for fortifiedWine in fortifiedWines {
+            //                    for preferred in preferredBases {
+            //                        if fortifiedWine.rawValue.lowercased() == preferred.name.lowercased() && cocktail != allCocktailsThatMatchBySpirit.last {
+            //                            allCocktailsThatMatchBySpirit.append(cocktail)
+            //                        }
+            //                    }
+            //
+            //                }
+            //            }
+            //            if let gins = cocktail.CompileTags().gin {
+            //                for gin in gins {
+            //                    for preferred in preferredBases {
+            //                        if gin.rawValue.lowercased() == preferred.name.lowercased() && cocktail != allCocktailsThatMatchBySpirit.last {
+            //                            allCocktailsThatMatchBySpirit.append(cocktail)
+            //                        }
+            //                    }
+            //
+            //                }
+            //            }
+            //            if let liqueurs = cocktail.CompileTags().liqueur {
+            //                for liqueur in liqueurs {
+            //                    for preferred in preferredBases {
+            //                        if liqueur.rawValue.lowercased() == preferred.name.lowercased() && cocktail != allCocktailsThatMatchBySpirit.last {
+            //                            allCocktailsThatMatchBySpirit.append(cocktail)
+            //                        }
+            //                    }
+            //
+            //                }
+            //            }
+            //            if let others = cocktail.CompileTags().other {
+            //                for other in others {
+            //                    for preferred in preferredBases {
+            //                        if other.rawValue.lowercased() == preferred.name.lowercased() && cocktail != allCocktailsThatMatchBySpirit.last {
+            //                            allCocktailsThatMatchBySpirit.append(cocktail)
+            //                        }
+            //                    }
+            //
+            //                }
+            //            }
+            //            if let rums = cocktail.CompileTags().rum {
+            //                for rum in rums {
+            //                    for preferred in preferredBases {
+            //                        if rum.rawValue.lowercased() == preferred.name.lowercased() && cocktail != allCocktailsThatMatchBySpirit.last {
+            //                            allCocktailsThatMatchBySpirit.append(cocktail)
+            //                        }
+            //                    }
+            //
+            //                }
+            //            }
+            //            if let vodkas = cocktail.CompileTags().vodka {
+            //                for vodka in vodkas {
+            //                    for preferred in preferredBases {
+            //                        if vodka.rawValue.lowercased() == preferred.name.lowercased() && cocktail != allCocktailsThatMatchBySpirit.last {
+            //                            allCocktailsThatMatchBySpirit.append(cocktail)
+            //                        }
+            //                    }
+            //
+            //                }
+            //            }
+            //            if let whiskies = cocktail.CompileTags().whiskies {
+            //                for whisky in whiskies {
+            //                    for preferred in preferredBases {
+            //                        if whisky.rawValue.lowercased() == preferred.name.lowercased() && cocktail != allCocktailsThatMatchBySpirit.last {
+            //                            allCocktailsThatMatchBySpirit.append(cocktail)
+            //                        }
+            //                    }
+            //
+            //                }
+            //            }
+            //            if let wines = cocktail.CompileTags().wine {
+            //                for wine in wines {
+            //                    for preferred in preferredBases {
+            //                        if wine.rawValue.lowercased() == preferred.name.lowercased() && cocktail != allCocktailsThatMatchBySpirit.last {
+            //                            allCocktailsThatMatchBySpirit.append(cocktail)
+            //                        }
+            //                    }
+            //
+            //                }
+            //            }
         }
-//        for i in allCocktailsThatMatchBySpirit {
-//            print(i.cocktailName)
-//        }
+        //        for i in allCocktailsThatMatchBySpirit {
+        //            print(i.cocktailName)
+        //        }
         
         // Take our new matchedBasesCocktails array, and rip out all the styles. Filter out the styles that match our PREFERREDSTYLESARRAY, and then add those to a new array called matchedStylesAndBasesCocktails.
         // matchedStylesAndBasesCocktails is therefor a subset of matchedBasesCocktails, that could be filtered
@@ -374,7 +373,7 @@ final class SearchCriteriaViewModel: ObservableObject {
         
         for matched in matchedProfilesTexturesBasesAndStylesCocktails {
             var internalMatchedCounter = 0
-
+            
             if let flavors = matched.CompileTags().flavors {
                 for flavor in flavors {
                     for preferredFlavor in preferredFlavors {
@@ -403,186 +402,186 @@ final class SearchCriteriaViewModel: ObservableObject {
         // if the base is in the unwanted array, remove it from the final array and update the matchedcount.
         for cocktail in matchedSet {
             
-//            if let agave = cocktail.CompileTags().agave {
-//                for agave in agave {
-//                    for unwanted in unwantedArray {
-//                        if agave.rawValue.lowercased() == unwanted.name.lowercased() {
-//                            matchedSet.remove(cocktail)
-//                            matchedCount -= 1
-//                        }
-//                    }
-//                    for preferred in preferredArray {
-//                        if agave.rawValue.lowercased() == preferred.name.lowercased() {
-//                            matchedCount += 1
-//                        }
-//                    }
-//                }
-//            }
-//            if let amari = cocktail.CompileTags().amari {
-//                for amaro in amari {
-//                    for unwanted in unwantedArray {
-//                        if amaro.rawValue.lowercased() == unwanted.name.lowercased() {
-//                            matchedSet.remove(cocktail)
-//                            matchedCount -= 1
-//                        }
-//                    }
-//                    for preferred in preferredArray {
-//                        if amaro.rawValue.lowercased() == preferred.name.lowercased() {
-//                            matchedCount += 1
-//                        }
-//                    }
-//                }
-//            }
-//            if let bitters = cocktail.CompileTags().bitters {
-//                for bitter in bitters {
-//                    for unwanted in unwantedArray {
-//                        if bitter.rawValue.lowercased() == unwanted.name.lowercased() {
-//                            matchedSet.remove(cocktail)
-//                            matchedCount -= 1
-//                        }
-//                    }
-//                    for preferred in preferredArray {
-//                        if bitter.rawValue.lowercased() == preferred.name.lowercased() {
-//                            matchedCount += 1
-//                        }
-//                    }
-//                }
-//            }
-//            if let brandies = cocktail.CompileTags().brandy {
-//                for brandy in brandies {
-//                    for unwanted in unwantedArray {
-//                        if brandy.rawValue.lowercased() == unwanted.name.lowercased() {
-//                            matchedSet.remove(cocktail)
-//                            matchedCount -= 1
-//                        }
-//                    }
-//                    for preferred in preferredArray {
-//                        if brandy.rawValue.lowercased() == preferred.name.lowercased() {
-//                            matchedCount += 1
-//                        }
-//                    }
-//                }
-//            }
-//            if let fortifiedWines = cocktail.CompileTags().fortifiedWine {
-//                for fortifiedWine in fortifiedWines {
-//                    for unwanted in unwantedArray {
-//                        if fortifiedWine.rawValue.lowercased() == unwanted.name.lowercased() {
-//                            matchedSet.remove(cocktail)
-//                            matchedCount -= 1
-//                        }
-//                    }
-//                    for preferred in preferredArray {
-//                        if fortifiedWine.rawValue.lowercased() == preferred.name.lowercased() {
-//                            matchedCount += 1
-//                        }
-//                    }
-//                }
-//            }
-//            if let gins = cocktail.CompileTags().gin {
-//                for gin in gins {
-//                    for unwanted in unwantedArray {
-//                        if gin.rawValue.lowercased() == unwanted.name.lowercased() {
-//                            matchedSet.remove(cocktail)
-//                            matchedCount -= 1
-//                        }
-//                    }
-//                    for preferred in preferredArray {
-//                        if gin.rawValue.lowercased() == preferred.name.lowercased() {
-//                            matchedCount += 1
-//                        }
-//                    }
-//                }
-//            }
-//            if let liqueurs = cocktail.CompileTags().liqueur {
-//                for liqueur in liqueurs {
-//                    for unwanted in unwantedArray {
-//                        if liqueur.rawValue.lowercased() == unwanted.name.lowercased() {
-//                            matchedSet.remove(cocktail)
-//                            matchedCount -= 1
-//                        }
-//                    }
-//                    for preferred in preferredArray {
-//                        if liqueur.rawValue.lowercased() == preferred.name.lowercased() {
-//                            matchedCount += 1
-//                        }
-//                    }
-//                }
-//            }
-//            if let others = cocktail.CompileTags().other {
-//                for other in others {
-//                    for unwanted in unwantedArray {
-//                        if other.rawValue.lowercased() == unwanted.name.lowercased() {
-//                            matchedSet.remove(cocktail)
-//                            matchedCount -= 1
-//                        }
-//                    }
-//                    for preferred in preferredArray {
-//                        if other.rawValue.lowercased() == preferred.name.lowercased() {
-//                            matchedCount += 1
-//                        }
-//                    }
-//                }
-//            }
-//            if let rums = cocktail.CompileTags().rum {
-//                for rum in rums {
-//                    for unwanted in unwantedArray {
-//                        if rum.rawValue.lowercased() == unwanted.name.lowercased() {
-//                            matchedSet.remove(cocktail)
-//                            matchedCount -= 1
-//                        }
-//                    }
-//                    for preferred in preferredArray {
-//                        if rum.rawValue.lowercased() == preferred.name.lowercased() {
-//                            matchedCount += 1
-//                        }
-//                    }
-//                }
-//            }
-//            if let vodkas = cocktail.CompileTags().vodka {
-//                for vodka in vodkas {
-//                    for unwanted in unwantedArray {
-//                        if vodka.rawValue.lowercased() == unwanted.name.lowercased() {
-//                            matchedSet.remove(cocktail)
-//                            matchedCount -= 1
-//                        }
-//                    }
-//                    for preferred in preferredArray {
-//                        if vodka.rawValue.lowercased() == preferred.name.lowercased() {
-//                            matchedCount += 1
-//                        }
-//                    }
-//                }
-//            }
-//            if let whiskies = cocktail.CompileTags().whiskies {
-//                for whisky in whiskies {
-//                    for unwanted in unwantedArray {
-//                        if whisky.rawValue.lowercased() == unwanted.name.lowercased() {
-//                            matchedSet.remove(cocktail)
-//                            matchedCount -= 1
-//                        }
-//                    }
-//                    for preferred in preferredArray {
-//                        if whisky.rawValue.lowercased() == preferred.name.lowercased() {
-//                            matchedCount += 1
-//                        }
-//                    }
-//                }
-//            }
-//            if let wines = cocktail.CompileTags().wine {
-//                for wine in wines {
-//                    for unwanted in unwantedArray {
-//                        if wine.rawValue.lowercased() == unwanted.name.lowercased() {
-//                            matchedSet.remove(cocktail)
-//                            matchedCount -= 1
-//                        }
-//                    }
-//                    for preferred in preferredArray {
-//                        if wine.rawValue.lowercased() == preferred.name.lowercased() {
-//                            matchedCount += 1
-//                        }
-//                    }
-//                }
-//            }
+            //            if let agave = cocktail.CompileTags().agave {
+            //                for agave in agave {
+            //                    for unwanted in unwantedArray {
+            //                        if agave.rawValue.lowercased() == unwanted.name.lowercased() {
+            //                            matchedSet.remove(cocktail)
+            //                            matchedCount -= 1
+            //                        }
+            //                    }
+            //                    for preferred in preferredArray {
+            //                        if agave.rawValue.lowercased() == preferred.name.lowercased() {
+            //                            matchedCount += 1
+            //                        }
+            //                    }
+            //                }
+            //            }
+            //            if let amari = cocktail.CompileTags().amari {
+            //                for amaro in amari {
+            //                    for unwanted in unwantedArray {
+            //                        if amaro.rawValue.lowercased() == unwanted.name.lowercased() {
+            //                            matchedSet.remove(cocktail)
+            //                            matchedCount -= 1
+            //                        }
+            //                    }
+            //                    for preferred in preferredArray {
+            //                        if amaro.rawValue.lowercased() == preferred.name.lowercased() {
+            //                            matchedCount += 1
+            //                        }
+            //                    }
+            //                }
+            //            }
+            //            if let bitters = cocktail.CompileTags().bitters {
+            //                for bitter in bitters {
+            //                    for unwanted in unwantedArray {
+            //                        if bitter.rawValue.lowercased() == unwanted.name.lowercased() {
+            //                            matchedSet.remove(cocktail)
+            //                            matchedCount -= 1
+            //                        }
+            //                    }
+            //                    for preferred in preferredArray {
+            //                        if bitter.rawValue.lowercased() == preferred.name.lowercased() {
+            //                            matchedCount += 1
+            //                        }
+            //                    }
+            //                }
+            //            }
+            //            if let brandies = cocktail.CompileTags().brandy {
+            //                for brandy in brandies {
+            //                    for unwanted in unwantedArray {
+            //                        if brandy.rawValue.lowercased() == unwanted.name.lowercased() {
+            //                            matchedSet.remove(cocktail)
+            //                            matchedCount -= 1
+            //                        }
+            //                    }
+            //                    for preferred in preferredArray {
+            //                        if brandy.rawValue.lowercased() == preferred.name.lowercased() {
+            //                            matchedCount += 1
+            //                        }
+            //                    }
+            //                }
+            //            }
+            //            if let fortifiedWines = cocktail.CompileTags().fortifiedWine {
+            //                for fortifiedWine in fortifiedWines {
+            //                    for unwanted in unwantedArray {
+            //                        if fortifiedWine.rawValue.lowercased() == unwanted.name.lowercased() {
+            //                            matchedSet.remove(cocktail)
+            //                            matchedCount -= 1
+            //                        }
+            //                    }
+            //                    for preferred in preferredArray {
+            //                        if fortifiedWine.rawValue.lowercased() == preferred.name.lowercased() {
+            //                            matchedCount += 1
+            //                        }
+            //                    }
+            //                }
+            //            }
+            //            if let gins = cocktail.CompileTags().gin {
+            //                for gin in gins {
+            //                    for unwanted in unwantedArray {
+            //                        if gin.rawValue.lowercased() == unwanted.name.lowercased() {
+            //                            matchedSet.remove(cocktail)
+            //                            matchedCount -= 1
+            //                        }
+            //                    }
+            //                    for preferred in preferredArray {
+            //                        if gin.rawValue.lowercased() == preferred.name.lowercased() {
+            //                            matchedCount += 1
+            //                        }
+            //                    }
+            //                }
+            //            }
+            //            if let liqueurs = cocktail.CompileTags().liqueur {
+            //                for liqueur in liqueurs {
+            //                    for unwanted in unwantedArray {
+            //                        if liqueur.rawValue.lowercased() == unwanted.name.lowercased() {
+            //                            matchedSet.remove(cocktail)
+            //                            matchedCount -= 1
+            //                        }
+            //                    }
+            //                    for preferred in preferredArray {
+            //                        if liqueur.rawValue.lowercased() == preferred.name.lowercased() {
+            //                            matchedCount += 1
+            //                        }
+            //                    }
+            //                }
+            //            }
+            //            if let others = cocktail.CompileTags().other {
+            //                for other in others {
+            //                    for unwanted in unwantedArray {
+            //                        if other.rawValue.lowercased() == unwanted.name.lowercased() {
+            //                            matchedSet.remove(cocktail)
+            //                            matchedCount -= 1
+            //                        }
+            //                    }
+            //                    for preferred in preferredArray {
+            //                        if other.rawValue.lowercased() == preferred.name.lowercased() {
+            //                            matchedCount += 1
+            //                        }
+            //                    }
+            //                }
+            //            }
+            //            if let rums = cocktail.CompileTags().rum {
+            //                for rum in rums {
+            //                    for unwanted in unwantedArray {
+            //                        if rum.rawValue.lowercased() == unwanted.name.lowercased() {
+            //                            matchedSet.remove(cocktail)
+            //                            matchedCount -= 1
+            //                        }
+            //                    }
+            //                    for preferred in preferredArray {
+            //                        if rum.rawValue.lowercased() == preferred.name.lowercased() {
+            //                            matchedCount += 1
+            //                        }
+            //                    }
+            //                }
+            //            }
+            //            if let vodkas = cocktail.CompileTags().vodka {
+            //                for vodka in vodkas {
+            //                    for unwanted in unwantedArray {
+            //                        if vodka.rawValue.lowercased() == unwanted.name.lowercased() {
+            //                            matchedSet.remove(cocktail)
+            //                            matchedCount -= 1
+            //                        }
+            //                    }
+            //                    for preferred in preferredArray {
+            //                        if vodka.rawValue.lowercased() == preferred.name.lowercased() {
+            //                            matchedCount += 1
+            //                        }
+            //                    }
+            //                }
+            //            }
+            //            if let whiskies = cocktail.CompileTags().whiskies {
+            //                for whisky in whiskies {
+            //                    for unwanted in unwantedArray {
+            //                        if whisky.rawValue.lowercased() == unwanted.name.lowercased() {
+            //                            matchedSet.remove(cocktail)
+            //                            matchedCount -= 1
+            //                        }
+            //                    }
+            //                    for preferred in preferredArray {
+            //                        if whisky.rawValue.lowercased() == preferred.name.lowercased() {
+            //                            matchedCount += 1
+            //                        }
+            //                    }
+            //                }
+            //            }
+            //            if let wines = cocktail.CompileTags().wine {
+            //                for wine in wines {
+            //                    for unwanted in unwantedArray {
+            //                        if wine.rawValue.lowercased() == unwanted.name.lowercased() {
+            //                            matchedSet.remove(cocktail)
+            //                            matchedCount -= 1
+            //                        }
+            //                    }
+            //                    for preferred in preferredArray {
+            //                        if wine.rawValue.lowercased() == preferred.name.lowercased() {
+            //                            matchedCount += 1
+            //                        }
+            //                    }
+            //                }
+            //            }
             
             // rip out and remove unwanted styles
             if let styles = cocktail.CompileTags().styles {
@@ -655,12 +654,12 @@ final class SearchCriteriaViewModel: ObservableObject {
                     
                 }
             }
-
+            
             // If the matchedCount is above the threshold (more than half of the preferredCount) we'll store it in an array with its matchedCount data (see MatchedCocktail struct)
             if matchedCount >= (preferredCount / 2) {
                 matchedCocktails.append(MatchedCocktail(cocktail: cocktail, count: preferredCount, matchedCount: matchedCount))
             }
-
+            
             // reset matchedCount for the next loop.
             matchedCount = 0
         }
@@ -668,7 +667,7 @@ final class SearchCriteriaViewModel: ObservableObject {
         
         // create sections for the Listview using the matchedCocktails array data we just gathered.
         for i in 0...Int(preferredCount / 2) {
-
+            
             let numberOfMatches = (preferredCount - i)
             // numberOfMatches will give us a the right section name. For example if preferredCount = 4 then numberOfMatches will be: (4-0 = 4 on the first loop, 4 - 1 = 3 on the second loop, 4 - 2 = 2 on the next loop, and then we stop because our cutoff is preferredCount / 2. Note that if preferredcount is an odd number, Int(count / 2) will round down which is why we're starting the loop at 0 instead of 1)
             
@@ -689,6 +688,7 @@ final class SearchCriteriaViewModel: ObservableObject {
         isLoading = false
     }
 }
+
 
 struct CocktailResultList: View {
     
@@ -718,7 +718,7 @@ struct CocktailResultList: View {
                                     } label: {
                                         Text(cocktail.cocktailName)
                                     }
-
+                                    
                                 }
                             }
                         }
@@ -738,103 +738,104 @@ struct CocktailResultList: View {
         }
     }
 }
-
-struct MatchedCocktail: Equatable {
-    let cocktail: Cocktail
-    let count: Int
-    let matchedCount: Int
-}
-
-struct ResultViewSectionData {
-    let id = UUID()
-    let count: Int
-    let matched: Int
-    let cocktails: [Cocktail]
-}
-
-class CocktailComponent: Identifiable, ObservableObject, Hashable {
     
-    static func == (lhs: CocktailComponent, rhs: CocktailComponent) -> Bool {
-        return lhs.id == rhs.id
+    
+    struct MatchedCocktail: Equatable {
+        let cocktail: Cocktail
+        let count: Int
+        let matchedCount: Int
     }
     
-    public func hash(into hasher: inout Hasher) {
-        return hasher.combine(id)
+    struct ResultViewSectionData {
+        let id = UUID()
+        let count: Int
+        let matched: Int
+        let cocktails: [Cocktail]
     }
     
-    
-    // I don't love this, but it will work for now..
-    
-    @Published var matchesCurrentSearch: Bool
-    var id = UUID()
-    var name: String
-    @Published var isPreferred: Bool = false
-    @Published var isUnwanted: Bool = false
-    var isSpirit: Bool = false
-    var isFlavor: Bool = false
-    var isProfile: Bool = false
-    var isStyle: Bool = false
-    var isTexture: Bool = false
-    var preferenceType: PreferenceType
-    var spiritCategory: BoozeCategory?
- 
-
-    init(name: String, isFlavor: Bool = false, isProfile: Bool = false, isStyle: Bool = false, isSpirit: Bool = false, isTexture: Bool = false, matchesCurrentSearch: Bool = true) {
-        self.name = name
-        self.isSpirit = isSpirit
-        self.isFlavor = isFlavor
-        self.isStyle = isStyle
-        self.isProfile = isProfile
-        self.isTexture = isTexture
-        self.matchesCurrentSearch = matchesCurrentSearch
-       
-        if isFlavor {
-            preferenceType = .flavors
-        } else if isStyle {
-            preferenceType = .style
-        } else if isSpirit {
-            preferenceType = .spirits
-        } else if isProfile {
-            preferenceType = .profiles
-        } else {
-            preferenceType = .textures
+    class CocktailComponent: Identifiable, ObservableObject, Hashable {
+        
+        // I don't love this, but it will work for now..
+        
+        @Published var matchesCurrentSearch: Bool
+        var id = UUID()
+        var name: String
+        @Published var isPreferred: Bool = false
+        @Published var isUnwanted: Bool = false
+        var isSpirit: Bool = false
+        var isFlavor: Bool = false
+        var isProfile: Bool = false
+        var isStyle: Bool = false
+        var isTexture: Bool = false
+        var preferenceType: PreferenceType
+        var spiritCategory: IngredientType
+        
+        
+        init(name: String, isFlavor: Bool = false, isProfile: Bool = false, isStyle: Bool = false, isSpirit: Bool = false, isTexture: Bool = false, matchesCurrentSearch: Bool = true) {
+            self.name = name
+            self.isSpirit = isSpirit
+            self.isFlavor = isFlavor
+            self.isStyle = isStyle
+            self.isProfile = isProfile
+            self.isTexture = isTexture
+            self.matchesCurrentSearch = matchesCurrentSearch
+            
+            if isFlavor {
+                preferenceType = .flavors
+            } else if isStyle {
+                preferenceType = .style
+            } else if isSpirit {
+                preferenceType = .spirits
+            } else if isProfile {
+                preferenceType = .profiles
+            } else {
+                preferenceType = .textures
+            }
         }
+        
+        init(for flavor: Flavor) {
+            self.name = flavor.rawValue
+            self.isFlavor = true
+            self.preferenceType = .flavors
+            self.matchesCurrentSearch = true
+        }
+        
+        init(for profile: Profile) {
+            self.name = profile.rawValue
+            self.isProfile = true
+            self.preferenceType = .profiles
+            self.matchesCurrentSearch = true
+        }
+        
+        init(for style: Style) {
+            self.name = style.rawValue
+            self.isStyle = true
+            self.preferenceType = .style
+            self.matchesCurrentSearch = true
+        }
+        
+        init(for texture: Texture) {
+            self.name = texture.rawValue
+            self.isTexture = true
+            self.preferenceType = .textures
+            self.matchesCurrentSearch = true
+        }
+        init(for booze: Booze) {
+            self.name = booze.ingredientType.name
+            self.isSpirit = true
+            self.preferenceType = .spirits
+            self.matchesCurrentSearch = true
+            self.spiritCategory = booze.ingredientType
+        }
+        
+        static func == (lhs: CocktailComponent, rhs: CocktailComponent) -> Bool {
+            return lhs.id == rhs.id
+        }
+        
+        public func hash(into hasher: inout Hasher) {
+            return hasher.combine(id)
+        }
+        
     }
     
-    init(for flavor: Flavor) {
-        self.name = flavor.rawValue
-        self.isFlavor = true
-        self.preferenceType = .flavors
-        self.matchesCurrentSearch = true
-    }
-
-    init(for profile: Profile) {
-        self.name = profile.rawValue
-        self.isProfile = true
-        self.preferenceType = .profiles
-        self.matchesCurrentSearch = true
-    }
     
-    init(for style: Style) {
-        self.name = style.rawValue
-        self.isStyle = true
-        self.preferenceType = .style
-        self.matchesCurrentSearch = true
-    }
-    init(for texture: Texture) {
-        self.name = texture.rawValue
-        self.isTexture = true
-        self.preferenceType = .textures
-        self.matchesCurrentSearch = true
-    }
-    init(for booze: Booze) {
-        self.name = booze.name
-        self.isSpirit = true
-        self.preferenceType = .spirits
-        self.matchesCurrentSearch = true
-        self.spiritCategory = booze.boozeCategory
-    }
-   
-}
-
-
